# openapi.json
{
    "components": {
        "schemas": {
...
            "CreatePerson": {
                "properties": {
                    "biography": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_by": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "date_of_birth": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "date_of_death": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "email": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "family_name": {
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "middle_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "nickname": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "place_of_birth": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "place_of_death": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "sex": {
                        "$ref": "#/components/schemas/Sex"
                    },
                    "tree": {
                        "description": "Name of the family tree this person belongs to. Must refer to an existing tree.",
                        "type": "string"
                    },
                    "username": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "family_name",
                    "first_name",
                    "sex",
                    "tree"
                ],
                "type": "object"
            },
            "ErrorResponse": {
                "description": "Error response body returned for all 4xx/5xx responses.",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
...
            "Person": {
                "properties": {
                    "biography": {
                        "description": "Short biography, up to 1000 characters.",
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_by": {
                        "description": "Identifier of the user who created this record.",
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "date_of_birth": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "date_of_death": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "email": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "family_name": {
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "id": {
                        "description": "Record ID in the form `person:<ulid>`.",
                        "example": "person:01jd4a8xyz",
                        "type": "string"
                    },
                    "image_path": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "middle_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "nickname": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "place_of_birth": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "place_of_death": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "sex": {
                        "$ref": "#/components/schemas/Sex"
                    },
                    "tree": {
                        "description": "Name of the family tree this person belongs to.",
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "username": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "verified": {
                        "description": "Whether this person's identity has been verified. Defaults to `false`.",
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "family_name",
                    "first_name",
                    "sex"
                ],
                "type": "object"
            },
...
            "Sex": {
                "enum": [
                    "Male",
                    "Female"
                ],
                "type": "string"
            },
...
            "UpdatePerson": {
                "description": "All fields are optional; only supplied fields are updated (MERGE semantics).",
                "properties": {
                    "biography": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_by": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "date_of_birth": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "date_of_death": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "email": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "family_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "first_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "middle_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "nickname": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "place_of_birth": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "place_of_death": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "sex": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/Sex"
                            }
                        ]
                    },
                    "username": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "verified": {
                        "type": [
                            "boolean",
                            "null"
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "REST API for ancestry and family tree data management, backed by SurrealDB.",
        "license": {
            "name": ""
        },
        "title": "clann-server API",
        "version": "0.1.0"
    },
    "openapi": "3.1.0",
    "paths": {
        "/api/persons": {
            "get": {
                "operationId": "list_persons",
                "parameters": [
                    {
                        "description": "When present, only return/allow access to persons with this `created_by` value.",
                        "in": "query",
                        "name": "created_by",
                        "required": false,
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ]
                        }
                    },
                    {
                        "description": "When present, only return persons belonging to this family tree.",
                        "in": "query",
                        "name": "tree",
                        "required": false,
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Person"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "List of persons, optionally filtered"
                    }
                },
                "tags": [
                    "persons"
                ]
            },
            "post": {
                "operationId": "create_person",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePerson"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        },
                        "description": "Person created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad request"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid request body"
                    }
                },
                "tags": [
                    "persons"
                ]
            }
        },
        "/api/persons/{id}": {
            "delete": {
                "operationId": "delete_person",
                "parameters": [
                    {
                        "description": "Person record ID (without the `person:` prefix)",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Person deleted"
                    }
                },
                "tags": [
                    "persons"
                ]
            },
            "get": {
                "operationId": "get_person",
                "parameters": [
                    {
                        "description": "Person record ID (without the `person:` prefix)",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        },
                        "description": "Person found"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Person not found"
                    }
                },
                "tags": [
                    "persons"
                ]
            },
            "put": {
                "operationId": "update_person",
                "parameters": [
                    {
                        "description": "Person record ID (without the `person:` prefix)",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePerson"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        },
                        "description": "Person updated"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Person not found"
                    }
                },
                "tags": [
                    "persons"
                ]
            }
        },
...
    },
...
}